I found that when you remove the child elements from the cards, then animating width, height, top, left, and shadow works fine. 60 fps on Safari and Chrome, and few framedrops on mobile Safari on iPhone 8, but okay.
This makes me think that a viable method to make the card expanding animations fast enough is to take screenshots of the content using the html2canvas library (if the library is fast and reliable enough) and then swapping out the card contents for the screenshots during animation.

The only alternative I can think of is to use transforms, but when we change the aspect ratio of the cards using a transform, then that stretches the borders and shadows, too, and I don't know how to work around that. Also it would force us to change the structure of the animations more since we need to 'counter-transform' the content to prevent it from stretching and do other weird stuff to position the content during the animation.
So therefore the other approach is preferable if html2canvas is fast and reliable enough.

Other stuff:
- One of the cards still drops frames on desktop safari. (The hello-orange-dark one) I suspect it's because the background image is too large, but not sure.